fix(switch_instances): more lost code
A bit more code I forgot to move over from the internal repo to the oss
one. This is from PR 19 on the old internal repo.

Did a `git diff internal/main..HEAD` and read through the diff looking
for red code that we wanted.
