+++ .env.example
@@ -26,8 +26,11 @@ MCP_TRANSPORT=stdio
 # These values can also be set/overridden in your MCP client config.
 AUTH_METHOD=native_app
 EVO_CLIENT_ID=Replace this with your client ID
-# The redirect URL where Bentley IMS sends the auth code after login.
+# The redirect URL where Bentley IMS sends the auth code after login (non-delegated mode only).
 # Must match a registered redirect URI in your IMS application.
+# In HTTP mode with CLIENT_DELEGATED_AUTH, the MCP server itself receives the
+# OAuth callback at {MCP_PUBLIC_BASE_URL}/auth/callback — register THAT URL in
+# your client instead. EVO_REDIRECT_URL is ignored in HTTP+delegated mode.
 EVO_REDIRECT_URL=http://localhost:3000/signin-callback
 # EVO_CLIENT_SECRET=your-client-secret  # Optional: only needed for client credentials flow
