+++ src/mcp_tools.py
@@ -24,7 +24,6 @@ The environment variables can be set in a .env file or
 passed directly to the MCP server as input parameters.
 """
 
-import json as _json
 import logging
 import os
 from pathlib import Path
@@ -32,9 +31,8 @@ from pathlib import Path
 from fastmcp import FastMCP
 from fastmcp.utilities.logging import configure_logging
 from starlette.middleware import Middleware
-from starlette.types import ASGIApp, Receive, Scope, Send
 
-from evo_mcp.client_auth import create_auth_provider
+from evo_mcp.client_auth import AuthMetadataPatchMiddleware, create_auth_provider
 from evo_mcp.tools import (
     register_admin_tools,
     register_file_tools,
@@ -356,97 +354,19 @@ if TOOL_FILTER in ["all", "data"]:
 
 if __name__ == "__main__":
     # Log startup information
-    logger = logging.getLogger(__name__)
+    from fastmcp.utilities.logging import get_logger
+    logger = get_logger(__name__)
     logger.info("Starting Evo MCP Server in %s mode", TRANSPORT.upper())
 
     # Run the server with selected transport mode
     if TRANSPORT == "http":
-
-        class AuthMetadataPatchMiddleware:
-            """Patches the OAuth metadata to advertise 'none' as a supported
-            token endpoint auth method, which is required for public MCP
-            clients (like VS Code) that don't have a client secret.
-
-            The MCP SDK hardcodes only ["client_secret_post", "client_secret_basic"],
-            but the OIDCProxy's DCR endpoint actually accepts "none".
-            """
-
-            def __init__(self, app: ASGIApp) -> None:
-                self.app = app
-
-            async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
-                if scope["type"] != "http":
-                    await self.app(scope, receive, send)
-                    return
-
-                path = scope.get("path", "")
-                method = scope.get("method", "?")
-                logger.debug("[req] %s %s", method, path)
-
-                if not path.startswith("/.well-known/oauth-authorization-server"):
-                    await self.app(scope, receive, send)
-                    return
-
-                # Capture response body, patch token_endpoint_auth_methods_supported
-                response_body = bytearray()
-                response_started = False
-                original_headers = []
-                original_status = 200
-
-                async def capture_send(message):
-                    nonlocal response_started, original_headers, original_status
-                    if message["type"] == "http.response.start":
-                        response_started = True
-                        original_status = message.get("status", 200)
-                        original_headers = list(message.get("headers", []))
-                    elif message["type"] == "http.response.body":
-                        response_body.extend(message.get("body", b""))
-
-                await self.app(scope, receive, capture_send)
-
-                try:
-                    data = _json.loads(bytes(response_body))
-                    methods = data.get("token_endpoint_auth_methods_supported", [])
-                    if "none" not in methods:
-                        methods.append("none")
-                        data["token_endpoint_auth_methods_supported"] = methods
-                    patched = _json.dumps(data).encode()
-
-                    # Update content-length header
-                    new_headers = [(k, v) for k, v in original_headers if k != b"content-length"]
-                    new_headers.append((b"content-length", str(len(patched)).encode()))
-
-                    await send(
-                        {
-                            "type": "http.response.start",
-                            "status": original_status,
-                            "headers": new_headers,
-                        }
-                    )
-                    await send(
-                        {
-                            "type": "http.response.body",
-                            "body": patched,
-                        }
-                    )
-                    logger.debug("Patched auth metadata: added 'none' to token_endpoint_auth_methods_supported")
-                except Exception:
-                    # If parsing fails, send original response unchanged
-                    await send(
-                        {
-                            "type": "http.response.start",
-                            "status": original_status,
-                            "headers": original_headers,
-                        }
-                    )
-                    await send(
-                        {
-                            "type": "http.response.body",
-                            "body": bytes(response_body),
-                        }
-                    )
-
         logger.info("HTTP server will listen on %s:%s", HTTP_HOST, HTTP_PORT)
+        if CLIENT_DELEGATED_AUTH:
+            logger.info(
+                "OAuth upstream callback URL: %s/auth/callback — "
+                "register this as an allowed redirect URI in your auth client.",
+                public_base_url.rstrip("/"),
+            )
         middleware = [Middleware(AuthMetadataPatchMiddleware)] if CLIENT_DELEGATED_AUTH else []
         mcp.run(
             transport="http",
@@ -454,6 +374,7 @@ if __name__ == "__main__":
             port=HTTP_PORT,
             middleware=middleware,
         )
+
     else:
         # Default STDIO mode
         mcp.run()
